Code cleanup: destructuring request data and error handling
A few request handlers only used the `url` and `method` properties
on the request to do their job. This can be made this clearer by
replacing lines like:

  app.use (req, res, next) ->

with:

  app.use ({url, method}, res, next) ->

The latter says "this function only cares about the url and method
properties of this argument".

Wherever we were handling requests like:

    return next() unless req.url is '/' and req.method is 'GET'

I've changed it to:

    unless url is '/' and method is 'GET' then do next ; return

To make it clear that we're calling next for its side-effects and
to make it clear that nothing is being returned. Also I think it's
slightly clearer (in this case) to have the condition before the
the action to take if the condition is false.